I think I fix
Experimental propagate settings check box
