--- widgets/__init__.py
@@ -1,14 +1,0 @@
-#
-# This file is part of Refinery.
-#
-# For authors and copyright check AUTHORS.TXT
-#
-# Refinery is free software under the GNU General Public License v3.0.
-# See LICENSE for more information.
-#
-
-from refinery.widgets import explorer_class_tree, explorer_hierarchy_tree,\
-     explorer_hybrid_tree, queue_tree
-
-__all__ = ("explorer_class_tree", "explorer_hierarchy_tree",
-           "explorer_hybrid_tree", "queue_tree")
