--- windows/__init__.py
@@ -1,14 +1,0 @@
-#
-# This file is part of Refinery.
-#
-# For authors and copyright check AUTHORS.TXT
-#
-# Refinery is free software under the GNU General Public License v3.0.
-# See LICENSE for more information.
-#
-
-from refinery.windows import actions_window, crc_window, meta_window,\
-     rename_window, settings_window
-
-__all__ = ("actions_window", "crc_window", "meta_window",
-           "rename_window", "settings_window")
